requirements.txt
aiohttp>=3.7.4.post0
aiounittest>=1.3.0
botbuilder>=0.0.1
botbuilder_azure>=4.14.1
datatypes_date_time>=1.0.0a2
msrest>=0.6.10
opencensus>=0.8.0
opencensus_ext_azure>=1.1.1
python_dateutil>=2.8.2
requests>=2.23.0
botbuilder.testing>=4.14.1
botbuilder.ai>=4.14.1
botbuilder-applicationinsights>=4.14.1
botbuilder-integration-applicationinsights-aiohttp>=4.14.1
